Modernize the client
Every part of the client is now under the assets/ directory and we use Parcel instead of Gulp to compile it as place the results to client/ directory. While at it, we stopped using jQuery in favour of vanilla JavaScript with jsx-dom for nicer templating.

The generated code expects everything relative to root so we had to move all static files there and make express serve them from /. This clashed with the `GET /` route so we had to move the staticServer after `GET /` handler.

We also needed to move penguin background to the client so that Parcel does not try to mangle its path.